.r Split Rust exec-guard git policy into a focused module (D5 partial)
Extract git control-plane policy from runtime/container/rust/src/lib.rs
into a new gitpolicy module: config-key/spec blocking, argument scanning
(should_block_reason), git-dir/work-tree path overrides, commit
short-flag no-verify detection, and ambient GIT_*/editor/pager env
override detection. Pure relocation — no behavior, ABI, or unsafe change.

- Exported C ABI symbol set unchanged: the 9 #[no_mangle]/extern "C"
  interposed symbols (execve/execv/execvp/execvpe/execveat/fexecve/
  posix_spawn/posix_spawnp/workcell_syscall_shim) are byte-identical
  before/after, verified via nm on the built cdylib.
- cargo build/test/clippy -D warnings/fmt --check all green (18 lib +
  12 bin tests). A4 stays clean: gitpolicy holds zero unsafe blocks;
  unsafe-extern guard passes. A3 fuzz_api public surface is unchanged
  and compiles under --cfg fuzzing.

Follow-on: the git-config blocklist moved with git_config_key_is_blocked
from lib.rs to gitpolicy.rs, so retarget the two repo-owned enforcers
that scanned lib.rs for that blocklist:
- internal/gitconfigblocklist (git-config-blocklist-parity invariant):
  enforcer path lib.rs -> gitpolicy.rs; parity now passes (all TOML keys
  + prefix/suffix patterns present in gitpolicy.rs).
- internal/mutation runner: core.fsmonitor mutation relativePath
  lib.rs -> gitpolicy.rs; empirically confirmed the mutation still
  applies and is killed by cargo test.
Also refreshed the two doc pointers (policy/git-config-blocklist.toml
and the enforcer docstring). These are enforcer metadata, not runtime
behavior — the split stays behavior/ABI-identical.

D5 is an L-sized item; this lands the cleanest, unsafe-free cut as one
reviewable PR. Follow-ups sequence the remaining tightly-coupled
surface: path-validation, runtime-protection, and shim modules.
